Fixed all type cast warnings

+++ Motor2D/Button.cpp
@@ -48,8 +48,8 @@ bool Button::PostUpdate()
 	}
 
 	ComputeRects();
-	rect.x = (-anchor_point.x * rect.w) + abs_pos.x;
-	rect.y = (-anchor_point.y * rect.h) + abs_pos.y;
+	rect.x = (int)(-anchor_point.x * rect.w) + abs_pos.x;
+	rect.y = (int)(-anchor_point.y * rect.h) + abs_pos.y;
 
 	SDL_Rect Mouse;
 	App->input->GetMousePosition(Mouse.x, Mouse.y);
