+++ Motor2D/j1Window.h
@@ -30,7 +30,7 @@ public:
 	void GetWindowSize(uint& width, uint& height) const;
 
 	// Retrieve window scale
-	uint GetScale() const;
+	float GetScale() const;
 
 public:
 	//The window we'll be rendering to
@@ -43,7 +43,7 @@ private:
 	p2SString	title;
 	uint		width;
 	uint		height;
-	uint		scale;
+	float		scale;
 };
 
 #endif // __j1WINDOW_H__
